+++ aim/web/ui/src/pages/RunDetail/TraceVisualizationContainer/RangePanel.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import RangeSliderWithInput from 'components/RangeSliderWithInput';
+import SliderWithInput from 'components/SliderWithInput';
 import { Button } from 'components/kit';
 
 interface IRangeSliderWithInputItem {
@@ -16,6 +16,7 @@ interface IRangeSliderWithInputItem {
   rangeEndpoints: [number, number];
   selectedRangeValue: [number, number];
   inputValue: number;
+  sliderType: 'single' | 'range'; // This type is same as SliderWithInput component sliderType prop type.
 }
 
 type RangeSliderData = IRangeSliderWithInputItem[];
@@ -45,9 +46,9 @@ function RangePanel({
     >
       <div className='RangePanelContainer'>
         {items?.map((item) => (
-          <>
-            <RangeSliderWithInput
-              key={item.sliderName}
+          <React.Fragment key={item.sliderName}>
+            <SliderWithInput
+              sliderType={item?.sliderType}
               sliderTitle={item.sliderTitle}
               countInputTitle={item.inputTitle}
               countTitleTooltip={item.inputTitleTooltip}
@@ -65,7 +66,7 @@ function RangePanel({
               }
             />
             <div className='VerticalDivider' />
-          </>
+          </React.Fragment>
         ))}
         <div className='ApplyButtonContainer'>
           <Button