+++ aim/web/ui/src/pages/RunDetail/TraceVisualizationContainer/TraceVisualizationContainer.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect } from 'react';
+import React, { memo, useEffect } from 'react';
 
 import Menu from 'components/kit/Menu/Menu';
 import BusyLoaderWrapper from 'components/BusyLoaderWrapper/BusyLoaderWrapper';
@@ -8,7 +8,8 @@ import useModel from 'hooks/model/useModel';
 import runTracesModel from 'services/models/runs/runTracesModel';
 
 import DistributionsVisualizer from '../DistributionsVisualizer';
-import ImagesVisualizer from '../ImagesVisualizer/ImagesVisualizer';
+import ImagesVisualizer from '../ImagesVisualizer';
+import PlotlyVisualizer from '../PlotlyVisualizer';
 import { ITraceVisualizationContainerProps } from '../types';
 
 import RangePanel from './RangePanel';
@@ -22,17 +23,23 @@ const traceTypeVisualization = {
   audios: () => null,
   videos: () => null,
   texts: () => null,
-  plotly: () => null,
+  figures: PlotlyVisualizer,
 };
 
 function TraceVisualizationContainer({
   traceInfo,
   traceType,
   runHash,
+  runParams,
 }: ITraceVisualizationContainerProps) {
   const runTracesModelData = useModel(runTracesModel);
   useEffect(() => {
-    runTracesModel.initialize(runHash, traceType, traceInfo[traceType]);
+    runTracesModel.initialize(
+      runHash,
+      traceType,
+      traceInfo[traceType],
+      runParams,
+    );
 
     return () => {
       runTracesModel.destroy();
@@ -57,7 +64,7 @@ function TraceVisualizationContainer({
       </div>
       <div className='VisualizerArea'>
         <BusyLoaderWrapper
-          height={'30rem'}
+          className='VisualizationLoader'
           isLoading={!!runTracesModelData?.isTraceBatchLoading}
         >
           <Visualizer
@@ -81,9 +88,10 @@ function TraceVisualizationContainer({
                 rangeEndpoints: runTracesModelData?.data[item.sliderName],
                 selectedRangeValue: runTracesModelData?.queryData?.sliders[
                   item.sliderName
-                ] || [0, 50],
+                ] || [0, 0],
                 inputValue:
-                  runTracesModelData?.queryData?.inputs[item.inputName] || 50,
+                  runTracesModelData?.queryData?.inputs[item.inputName] || 0,
+                sliderType: item.sliderType,
               }))}
               onApply={runTracesModel.onApply}
               onInputChange={runTracesModel.onInputChange}
@@ -98,4 +106,4 @@ function TraceVisualizationContainer({
 
 TraceVisualizationContainer.displayName = 'TraceVisualizationContainer';
 
-export default widthEmptyTraceCheck(TraceVisualizationContainer);
+export default memo(widthEmptyTraceCheck(TraceVisualizationContainer));