+++ aim/web/ui/src/services/api/runs/runsService.ts
@@ -1,5 +1,6 @@
+import { TraceType } from 'services/models/runs/types';
+
 import API from '../api';
-import { TraceRawDataItem, TraceType } from '../../models/runs/types';
 
 const endpoints = {
   GET_RUNS: 'runs/search/run',