+++ aim/web/ui/src/services/models/runs/settings.ts
@@ -1,4 +1,10 @@
-import { processDistributionsData, processImagesData } from './util';
+import { QueryData } from './types';
+import {
+  processDistributionsData,
+  processImagesData,
+  processPlotlyData,
+  reformatArrayQueries,
+} from './util';
 
 type InputItem = {
   title: string;
@@ -10,12 +16,14 @@ type SliderItem = {
   title: string;
   defaultValue: [number, number];
   tooltip: string;
+  sliderType: 'single' | 'range'; // This type is same as SliderWithInput component sliderType prop type.
 };
 
 type SettingItem = {
   dataProcessor: Function;
   sliders: Record<string, SliderItem>;
   inputs: Record<string, InputItem>;
+  paramsToApi?: (queryData?: QueryData) => Record<string, unknown>;
 };
 
 const settings: Record<string, SettingItem> = {
@@ -26,6 +34,7 @@ const settings: Record<string, SettingItem> = {
         defaultValue: [0, 50],
         title: 'Steps',
         tooltip: 'Training step. Increments every time track() is called',
+        sliderType: 'range',
       },
     },
     inputs: {
@@ -41,25 +50,55 @@ const settings: Record<string, SettingItem> = {
     sliders: {
       record_range: {
         defaultValue: [0, 50],
-        tooltip: '',
-        title: '',
+        tooltip: 'Training step. Increments every time track() is called',
+        title: 'Steps',
+        sliderType: 'range',
       },
       index_range: {
         defaultValue: [0, 50],
-        tooltip: '',
-        title: '',
+        tooltip: 'Index in the list of images passed to track() call',
+        title: 'Indices',
+        sliderType: 'range',
       },
     },
     inputs: {
       record_density: {
         defaultValue: 50,
-        title: '',
-        tooltip: '',
+        title: 'Steps count',
+        tooltip: 'Number of steps to display',
       },
       index_density: {
         defaultValue: 5,
-        title: '',
-        tooltip: '',
+        title: 'Indices count',
+        tooltip: 'Number of images per step',
+      },
+    },
+  },
+  figures: {
+    dataProcessor: processPlotlyData,
+    paramsToApi: (queryData?: QueryData) => {
+      const record_range = queryData?.inputs?.record_range || 1;
+
+      return {
+        ...reformatArrayQueries({
+          record_range: [record_range - 1, record_range],
+        }),
+        record_density: 1,
+      };
+    },
+    sliders: {
+      record_range: {
+        defaultValue: [0, 0],
+        tooltip: 'Training step. Increments every time track() is called',
+        title: 'Step',
+        sliderType: 'single',
+      },
+    },
+    inputs: {
+      record_range: {
+        defaultValue: 1,
+        tooltip: 'Training step. To see Plotly tracked in the step.',
+        title: 'Step',
       },
     },
   },