+++ aim/web/ui/src/utils/app/onSortFieldsChange.ts
@@ -1,4 +1,4 @@
-import { isEmpty } from 'lodash';
+import { isEmpty } from 'lodash-es';
 
 import * as analytics from 'services/analytics';