+++ aim/web/ui/src/utils/app/resetChartZoom.ts
@@ -1,22 +1,27 @@
 import * as analytics from 'services/analytics';
 
-import { IMetricAppConfig } from 'types/services/models/metrics/metricsAppModel';
+import {
+  IAppModelConfig,
+  ILineChartConfig,
+} from 'types/services/models/explorer/createAppModel';
 
 export default function resetChartZoom({
   configData,
   appName,
 }: {
-  configData: IMetricAppConfig | any;
+  configData: IAppModelConfig;
   appName: string;
-}): IMetricAppConfig | any {
-  configData.chart = {
-    ...configData.chart,
-    zoom: {
-      ...configData.chart.zoom,
-      active: false,
-      history: [],
-    },
-  };
+}): IAppModelConfig {
+  if (configData.chart) {
+    configData.chart = {
+      ...configData.chart,
+      zoom: {
+        ...(configData.chart as ILineChartConfig).zoom,
+        active: false,
+        history: [],
+      },
+    };
+  }
   analytics.trackEvent(`[${appName}Explorer][Chart] Reset zoom`);
   return configData;
 }