+++ aim/web/ui/src/utils/d3/lineGenerator.ts
@@ -1,12 +1,12 @@
 import * as d3 from 'd3';
 
-import { IGetAxisScale } from 'types/utils/d3/getAxisScale';
+import { IAxisScale } from 'types/utils/d3/getAxisScale';
 
 import { CurveEnum } from './';
 
 export default function lineGenerator(
-  xScaleValues: IGetAxisScale,
-  yScaleValues: IGetAxisScale | { [key: string]: IGetAxisScale },
+  xScaleValues: IAxisScale,
+  yScaleValues: IAxisScale | { [key: string]: IAxisScale },
   curve: CurveEnum = CurveEnum.Linear,
 ) {
   if (!xScaleValues) {